Tests of Unenclosed JSON: rename so MRv2 is more prominent, MRv1 as legacy (consistent with TestUnenclosedGeoJsonRecordReader which is MRv2)
